+++ tests/Integration/SourceAdapters/SourceAdapterTest.php
@@ -19,24 +19,11 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class SourceAdapterTest extends TestCase
 {
-    private const READABLE_MODES = [
-        'r' => true,
-        'w+' => true,
-        'r+' => true,
-        'x+' => true,
-        'c+' => true,
-        'rb' => true,
-        'w+b' => true,
-        'r+b' => true,
-        'x+b' => true,
-        'c+b' => true,
-        'rt' => true,
-        'w+t' => true,
-        'r+t' => true,
-        'x+t' => true,
-        'c+t' => true,
-        'a+' => true
-    ];
+    private const EXPECTED_FILENAME = 'plank';
+    private const EXPECTED_EXTENSION = 'png';
+    private const EXPECTED_MIME = 'image/png';
+    private const EXPECTED_SIZE = 7173;
+    private const EXPECTED_HASH = '3ef5e70366086147c2695325d79a25cc';
 
     public function setUp(): void
     {
@@ -79,119 +66,174 @@ class SourceAdapterTest extends TestCase
         $memoryStream->write($string);
 
         $dataStream = Utils::streamFor(fopen('data://image/png,' . rawurlencode($string), 'rb'));
-        $mime = 'image/png';
         $data = [
             'FileAdapter' => [
                 'adapterClass' => FileAdapter::class,
                 'source' => new File($file),
                 'path' => $file,
-                'filename' => 'plank',
-                'inferredMime' => $mime,
-                'clientMime' => null
+                'filename' => self::EXPECTED_FILENAME,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => null,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'UploadedFileAdapter' => [
                 'adapterClass' => UploadedFileAdapter::class,
                 'source' => $uploadedFile,
                 'path' => $file,
-                'filename' => 'plank',
-                'inferredMime' => $mime,
-                'clientMime' => $mime
+                'filename' => self::EXPECTED_FILENAME,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => self::EXPECTED_MIME,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'LocalPathAdapter' => [
                 'adapterClass' => LocalPathAdapter::class,
                 'source' => $file,
                 'path' => $file,
-                'filename' => 'plank',
-                'inferredMime' => $mime,
-                'clientMime' => null
+                'filename' => self::EXPECTED_FILENAME,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => null,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'RemoteUrlAdapter' => [
                 'adapterClass' => RemoteUrlAdapter::class,
                 'source' => $url,
                 'path' => $url,
-                'filename' => 'plank',
-                'inferredMime' => null,
-                'clientMime' => $mime
+                'filename' => self::EXPECTED_FILENAME,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => self::EXPECTED_MIME,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'RawContentAdapter' => [
                 'adapterClass' => RawContentAdapter::class,
                 'source' => $string,
                 'path' => null,
-                'filename' => '',
-                'inferredMime' => $mime,
-                'clientMime' => null
+                'filename' => null,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => null,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'DataUrlAdapter_base64' => [
                 'adapterClass' => DataUrlAdapter::class,
                 'source' => $base64DataUrl,
                 'path' => null,
-                'filename' => '',
-                'inferredMime' => $mime,
-                'clientMime' => $mime
+                'filename' => null,
+                'extension' => null,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => self::EXPECTED_MIME,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'DataUrlAdapter_urlencode' => [
                 'adapterClass' => DataUrlAdapter::class,
                 'source' => $rawDataUrl,
                 'path' => null,
-                'filename' => '',
-                'inferredMime' => $mime,
-                'clientMime' => $mime
+                'filename' => null,
+                'extension' => null,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => self::EXPECTED_MIME,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'StreamResourceAdapter_Local' => [
                 'adapterClass' => StreamResourceAdapter::class,
                 'source' => $fileResource,
                 'path' => $file,
-                'filename' => 'plank',
-                'inferredMime' => $mime,
-                'clientMime' => null
+                'filename' => self::EXPECTED_FILENAME,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => null,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'StreamAdapter_Local' => [
                 'adapterClass' => StreamAdapter::class,
                 'source' => $fileStream,
                 'path' => $file,
-                'filename' => 'plank',
-                'inferredMime' => $mime,
-                'clientMime' => null
+                'filename' => self::EXPECTED_FILENAME,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => null,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'StreamResourceAdapter_Remote' => [
                 'adapterClass' => StreamResourceAdapter::class,
                 'source' => $httpResource,
                 'path' => $url,
-                'filename' => 'plank',
-                'inferredMime' => null,
-                'clientMime' => $mime
+                'filename' => self::EXPECTED_FILENAME,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => self::EXPECTED_MIME,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'StreamAdapter_Remote' => [
                 'adapterClass' => StreamAdapter::class,
                 'source' => $httpStream,
                 'path' => $url,
-                'filename' => 'plank',
-                'inferredMime' => null,
-                'clientMime' => $mime
+                'filename' => self::EXPECTED_FILENAME,
+                'extension' => self::EXPECTED_EXTENSION,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => self::EXPECTED_MIME,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'StreamResourceAdapter_Memory' => [
                 'adapterClass' => StreamResourceAdapter::class,
                 'source' => $memoryResource,
-                'path' => '',
-                'filename' => '',
-                'inferredMime' => $mime,
-                'clientMime' => null
+                'path' => null,
+                'filename' => null,
+                'extension' => null,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => null,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'StreamAdapter_Memory' => [
                 'adapterClass' => StreamAdapter::class,
                 'source' => $memoryStream,
-                'path' => '',
-                'filename' => '',
-                'inferredMime' => $mime,
-                'clientMime' => null
+                'path' => null,
+                'filename' => null,
+                'extension' => null,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => null,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
             'StreamAdapter_Data' => [
                 'adapterClass' => StreamAdapter::class,
                 'source' => $dataStream,
-                'path' => '',
-                'filename' => '',
-                'inferredMime' => $mime,
-                'clientMime' => $mime
+                'path' => null,
+                'filename' => null,
+                'extension' => null,
+                'inferredMime' => self::EXPECTED_MIME,
+                'clientMime' => self::EXPECTED_MIME,
+                'isRemote' => false,
+                'size' => self::EXPECTED_SIZE,
+                'hash' => self::EXPECTED_HASH,
             ],
         ];
         return $data;
@@ -224,97 +266,33 @@ class SourceAdapterTest extends TestCase
     /**
      * @dataProvider adapterProvider
      */
-    public function test_it_can_return_source($adapterClass, $source)
-    {
-        /** @var SourceAdapterInterface $adapter */
-        $adapter = new $adapterClass($source);
-        $this->assertEquals($source, $adapter->getSource());
-    }
-
-    /**
-     * @dataProvider adapterProvider
-     */
-    public function test_it_adapts_absolute_path($adapterClass, $source, $path)
-    {
-        /** @var SourceAdapterInterface $adapter */
-        $adapter = new $adapterClass($source);
-        $this->assertEquals($path, $adapter->path());
-    }
-
-    /**
-     * @dataProvider adapterProvider
-     */
-    public function test_it_adapts_filename($adapterClass, $source, $path, $filename)
-    {
-        /** @var SourceAdapterInterface $adapter */
-        $adapter = new $adapterClass($source);
-        $this->assertSame($filename, $adapter->filename());
-    }
-
-    /**
-     * @dataProvider adapterProvider
-     */
-    public function test_it_adapts_extension($adapterClass, $source)
-    {
-        /** @var SourceAdapterInterface $adapter */
-        $adapter = new $adapterClass($source);
-        $this->assertEquals('png', $adapter->extension());
-    }
-
-    /**
-     * @dataProvider adapterProvider
-     */
-    public function test_it_adapts_mime_type($adapterClass, $source, $path, $filename, $inferredMime, $clientMime)
-    {
-        /** @var SourceAdapterInterface $adapter */
-        $adapter = new $adapterClass($source);
-        $this->assertEquals($inferredMime, $adapter->mimeType());
-        $this->assertEquals($clientMime, $adapter->clientMimeType());
-    }
-
-    /**
-     * @dataProvider adapterProvider
-     */
-    public function test_it_adapts_file_contents($adapterClass, $source)
-    {
-        /** @var SourceAdapterInterface $adapter */
-        $adapter = new $adapterClass($source);
-        $this->assertIsString($adapter->contents());
-    }
-
-    /**
-     * @dataProvider adapterProvider
-     */
-    public function test_it_adapts_to_stream(
-        $adapterClass,
-        $source
+    public function test_it_extracts_expected_information_from_source(
+        string $adapterClass,
+        mixed $source,
+        ?string $path,
+        ?string $filename,
+        ?string $extension,
+        ?string $inferredMime,
+        ?string $clientMime,
+        ?bool $isRemote,
+        ?int $size,
+        ?string $hash
     ) {
         /** @var SourceAdapterInterface $adapter */
         $adapter = new $adapterClass($source);
-        $stream = $adapter->getStream();
+        $this->assertTrue($adapter->valid());
 
+        $stream = $adapter->getStream();
         $this->assertInstanceOf(StreamInterface::class, $stream);
         $this->assertTrue($stream->isReadable());
-    }
 
-    /**
-     * @dataProvider adapterProvider
-     */
-    public function test_it_adapts_file_size($adapterClass, $source)
-    {
-        /** @var SourceAdapterInterface $adapter */
-        $adapter = new $adapterClass($source);
-        $this->assertEquals(7173, $adapter->size());
-    }
-
-    /**
-     * @dataProvider adapterProvider
-     */
-    public function test_it_verifies_file_validity($adapterClass, $source)
-    {
-        /** @var SourceAdapterInterface $adapter */
-        $adapter = new $adapterClass($source);
-        $this->assertTrue($adapter->valid());
+        $this->assertSame($path, $adapter->path());
+        $this->assertSame($filename, $adapter->filename());
+        $this->assertSame($extension, $adapter->extension());
+        $this->assertSame($inferredMime, $adapter->mimeType());
+        $this->assertSame($clientMime, $adapter->clientMimeType());
+        $this->assertSame($size, $adapter->size());
+        $this->assertSame($hash, $adapter->hash());
     }
 
     /**